Co-locate Expo Android view managers with their native views
