+++ packages/expo/android/src/main/java/expo/modules/clerk/ClerkUserProfileExpoView.kt
@@ -11,7 +11,12 @@ import com.clerk.api.Clerk
 import com.clerk.ui.userprofile.UserProfileView
 import com.facebook.react.bridge.Arguments
 import com.facebook.react.bridge.ReactContext
+import com.facebook.react.common.MapBuilder
+import com.facebook.react.uimanager.SimpleViewManager
+import com.facebook.react.uimanager.ThemedReactContext
 import com.facebook.react.uimanager.events.RCTEventEmitter
+import com.facebook.react.uimanager.annotations.ReactProp
+import com.facebook.react.viewmanagers.ClerkUserProfileViewManagerInterface
 
 private const val TAG = "ClerkUserProfileExpoView"
 
@@ -46,5 +51,30 @@ class ClerkUserProfileNativeView(context: Context) : ClerkComposeNativeViewHost(
     }
     reactContext.getJSModule(RCTEventEmitter::class.java)
       .receiveEvent(id, "onProfileEvent", eventData)
+  }
+}
+
+class ClerkUserProfileViewManager : SimpleViewManager<ClerkUserProfileNativeView>(),
+    ClerkUserProfileViewManagerInterface<ClerkUserProfileNativeView> {
+
+  override fun getName(): String = "ClerkUserProfileView"
+
+  override fun createViewInstance(reactContext: ThemedReactContext): ClerkUserProfileNativeView {
+    return ClerkUserProfileNativeView(reactContext)
+  }
+
+  @ReactProp(name = "isDismissible")
+  override fun setIsDismissible(view: ClerkUserProfileNativeView, isDismissible: Boolean) {
+    view.isDismissible = isDismissible
+    view.setupView()
+  }
+
+  override fun getExportedCustomBubblingEventTypeConstants(): MutableMap<String, Any>? {
+    return MapBuilder.builder<String, Any>()
+      .put("onProfileEvent", MapBuilder.of(
+        "phasedRegistrationNames",
+        MapBuilder.of("bubbled", "onProfileEvent")
+      ))
+      .build() as MutableMap<String, Any>
   }
 }
